+++ CHANGES.rst
@@ -5,6 +5,7 @@ Changes
 ----------------
 
 - Rewrite edit link on server to navigate to Tekir admin page.
+- Add support for publishing site to servers.
 - Add support for adding attachments.
 - Add support for opening relevant folders in file manager.
 - Make UI language preference persistent.